Move tag listing classes into main library (instead of examples)
